Add logic to handle message fragmentation
Collectd had a bad habit of fragmenting messages. In short, it will
send two POSTS that are meant to be handled together. In some cases,
these POSTS will each contain part of a single line of data.

This logic stores fragments for the next POST to put back together.
It discards the fragment from the first POST because it sometimes is
complete enough to parse and really mess up the DB.

Fragmentation seems to be discoverable by looking for a lack of blank
line at the end of the first POST.

Example fragments:
First POST, last line: PUTVAL host.example.com/disk-d
Second POST, first line: m-0/disk_octets interval=10 1314724861:654568